% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createNeoWordNet.R
\name{createNeoWordNet}
\alias{createNeoWordNet}
\title{Create a WordNet graph database in Neo4j}
\usage{
createNeoWordNet(dictPath = paste(getwd(), "dict", sep = "/"),
  verbose = TRUE, url = "http://localhost:7474/db/data/",
  username = "neo4j", password = "graph")
}
\arguments{
\item{dictPath}{A path to the WordNet data and index files. Defaults to working directory.}

\item{verbose}{Should the function provide progress details? Defaults to true.}

\item{url}{Location of (running) Neo4j graph database. Defaults to local host.}

\item{username}{Username for Neo4j graph database access. Defaults to "neo4j"}

\item{password}{Password for Neo4j graph database access. Defaults to "graph"}
}
\description{
\code{createNeoWordNet} processes WordNet index and data files and returns a
graph object for further analysis. This requires approximately 425 MB of
storage and can take several hours to complete. (Approximately 6 hours on a
2 GHz Intel Core 2 Duo MacBook with 4 GB of 1067 MHz DDR3 memory)
}

