% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadWordNetFiles.R
\name{downloadWordNetFiles}
\alias{downloadWordNetFiles}
\title{Download WordNet database files}
\usage{

  downloadWordNetFiles(wordNetUrl = "http://wordnetcode.princeton.edu/wn3.1.dict.tar.gz",
  dest = getwd(), verbose = TRUE)
}
\arguments{
\item{wordNetUrl}{URL for the desired version of the WordNet database files.
Defaults to version 3.1.}

\item{dest}{Destination path for WordNet files.}

\item{verbose}{Should the function provide progress details? Defaults to
true.}
}
\description{
\code{downloadWordNetFiles} downloads and processes WordNet index and
data files for later load into Neo4j
}

