% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAndCreateNeoWN.R
\name{getAndCreateNeoWN}
\alias{getAndCreateNeoWN}
\title{Download WordNet files and create a WordNet graph database in single command}
\usage{

  getAndCreateNeoWN(wordNetUrl = "http://wordnetcode.princeton.edu/wn3.1.dict.tar.gz",
  dest = getwd(), url = "http://localhost:7474/db/data/",
  username = "neo4j", password = "graph", verbose = TRUE)
}
\arguments{
\item{wordNetUrl}{URL for the desired version of the WordNet database files.
Defaults to version 3.1.}

\item{dest}{A path to the WordNet data and index files. Defaults to
working directory.}

\item{url}{Location of (running) Neo4j graph database. Defaults to local
host.}

\item{username}{Username for Neo4j graph database access. Defaults to 'neo4j'}

\item{password}{Password for Neo4j graph database access. Defaults to 'graph'}

\item{verbose}{Should the function provide progress details? Defaults to
true.}
}
\description{
\code{getAndCreateNeoWN} downloads and processes WordNet index and
data files and returns a graph object for further analysis. This requires
approximately 425 MB of storage and can take several hours to complete.
(Approximately 6 hours on a 2 GHz Intel Core 2 Duo MacBook with 4 GB of 1067
MHz DDR3 memory)
}

